requirements.txt
streamlit
pillow
scikit-learn
pandas
numpy
python-dotenv
torch
sentence-transformers
matplotlib
google-generativeai